Only template mapping functions.
