+++ src/paired_end_mapping_metadata.h
@@ -88,7 +88,6 @@ class PairedEndMappingMetadata {
   friend class CandidateProcessor;
   template <typename MappingRecord>
   friend class MappingGenerator;
-  template <typename MappingRecord>
   friend class Chromap;
 };
